I removed the map implemetation, since I don't think it is as challenging to implemented, since it is just a Set of Pairs.
I added a way for the user to define it's own hash function as the primary hash. it is type, that defaults to std::hash<T>.

I added a method in main.cpp to insert elements that will cause collosion, so that the user can see the different way of handling it based on probing method.

a sample problem of finding duplicateChar in a string is added, with 3 different solutions, taking time from O(N^2) to O(N), that we can solve efficiently using Set.

Signed-off-by: Hassan Attar <hattar00@mylangara.ca>