Added UniversalHash class, returning a member of Universal Hash family of functions randomly
This struct is capable of returning 999,965,000,306 different hash functions
that will distribute the keys uniformely in the hash table. the probability of colision is upper bounded by O(1/n)
n being the total number of slots in the hash table if the maximum keys being inserted into the hashTable is chosen carefully.

by modifying the UniversalHash class to choose a prime, greater than the maximum key, this property will hold:
Pr(colision) = O(1/n)

Signed-off-by: Hassan Attar <hattar00@mylangara.ca>